pyproject.toml
[project]
name = "gh-chess-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygithub>=2.4.0",
    "requests>=2.32.3",
]
